Test the needed modules
Dropped the use of Graphics:: modules. Color::Rgb supplies all the tools which I
believe will be needed, so far. The list of names in Graphics:: didn't include
RebeccaPurple and Color::Rgb does.

Coded some samples and tests for how to use the module. Keeping them in, for the
time, as a reminder of how to use the included module. HSL color syntax is a
possible upgrade later, mostly as I might need that system for some or what I'm
trying to accomplish here anyway.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>