Add extensive demo to the code
The demo attempts to "exercise" almost every option for conversion between the
available formats. Exempted are the trivial conversions between string and tuple
versions of the same color and format method. Converting a 3-elemnt tuple to a
string requires no functionality not included in the Perl core, so there's not
a reason to write a routine to handle such.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>