Add rgb to hex routine
The rgb2hex routine in the Color::Rgb module does simple integer truncation of
the tuple vaules. For values of integers this is not a problem, but for the
results of calculations, where values only come close, at best, the errors were
unacceptable. For example, #121212 with a 100% overlay of #ffffff resulted in
a value of #fefeff. Not!

Opted for a simple process, with range checking and returning a value of zero
for out-of-range elements. As the Math::Round package is already needed, just
used that for the rounding.

The name to rgb or hex and the hex to rgb should still be fine. The hex values
are, by definition, integers already, and the named colors have integer values
assigned to the. Future development might have me replace those functions also,
but not at the moment.

Moved descriptions into the subroutines. Makes my screen cleaner when I collapse
functions and sections. Adjusted the formatting of opening comments and a few
other places as well. All without any substantive changes.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>